+++ source4/dns_server/dns_utils.c
@@ -33,79 +33,6 @@
 #undef DBGC_CLASS
 #define DBGC_CLASS DBGC_DNS
 
-/* Names are equal if they match and there's nothing left over */
-bool dns_name_equal(const char *name1, const char *name2)
-{
-	size_t host_part_len;
-	bool ret = dns_name_match(name1, name2, &host_part_len);
-
-	return ret && (host_part_len == 0);
-}
-
-/*
-  see if two dns records match
- */
-bool dns_records_match(struct dnsp_DnssrvRpcRecord *rec1,
-		       struct dnsp_DnssrvRpcRecord *rec2)
-{
-	bool status;
-	int i;
-
-	if (rec1->wType != rec2->wType) {
-		return false;
-	}
-
-	/* see if the data matches */
-	switch (rec1->wType) {
-	case DNS_TYPE_A:
-		return strcmp(rec1->data.ipv4, rec2->data.ipv4) == 0;
-	case DNS_TYPE_AAAA:
-		return strcmp(rec1->data.ipv6, rec2->data.ipv6) == 0;
-	case DNS_TYPE_CNAME:
-		return dns_name_equal(rec1->data.cname, rec2->data.cname);
-	case DNS_TYPE_TXT:
-		if (rec1->data.txt.count != rec2->data.txt.count) {
-			return false;
-		}
-		status = true;
-		for (i=0; i<rec1->data.txt.count; i++) {
-			status = status && (strcmp(rec1->data.txt.str[i],
-						rec2->data.txt.str[i]) == 0);
-		}
-		return status;
-	case DNS_TYPE_PTR:
-		return strcmp(rec1->data.ptr, rec2->data.ptr) == 0;
-	case DNS_TYPE_NS:
-		return dns_name_equal(rec1->data.ns, rec2->data.ns);
-
-	case DNS_TYPE_SRV:
-		return rec1->data.srv.wPriority == rec2->data.srv.wPriority &&
-			rec1->data.srv.wWeight  == rec2->data.srv.wWeight &&
-			rec1->data.srv.wPort    == rec2->data.srv.wPort &&
-			dns_name_equal(rec1->data.srv.nameTarget, rec2->data.srv.nameTarget);
-
-	case DNS_TYPE_MX:
-		return rec1->data.mx.wPriority == rec2->data.mx.wPriority &&
-			dns_name_equal(rec1->data.mx.nameTarget, rec2->data.mx.nameTarget);
-
-	case DNS_TYPE_HINFO:
-		return strcmp(rec1->data.hinfo.cpu, rec2->data.hinfo.cpu) == 0 &&
-			strcmp(rec1->data.hinfo.os, rec2->data.hinfo.os) == 0;
-
-	case DNS_TYPE_SOA:
-		return dns_name_equal(rec1->data.soa.mname, rec2->data.soa.mname) &&
-			dns_name_equal(rec1->data.soa.rname, rec2->data.soa.rname) &&
-			rec1->data.soa.serial == rec2->data.soa.serial &&
-			rec1->data.soa.refresh == rec2->data.soa.refresh &&
-			rec1->data.soa.retry == rec2->data.soa.retry &&
-			rec1->data.soa.expire == rec2->data.soa.expire &&
-			rec1->data.soa.minimum == rec2->data.soa.minimum;
-	default:
-		break;
-	}
-
-	return false;
-}
 
 /*
  * Lookup a DNS record, performing an exact match.
