+++ source4/dns_server/dnsserver_common.h
@@ -78,6 +78,9 @@ WERROR dns_common_name2dn(struct ldb_context *samdb,
 			  struct ldb_dn **_dn);
 bool dns_name_equal(const char *name1, const char *name2);
 
+bool dns_records_match(struct dnsp_DnssrvRpcRecord *rec1,
+		       struct dnsp_DnssrvRpcRecord *rec2);
+
 /*
  * For this routine, base_dn is generally NULL.  The exception comes
  * from the python bindings to support setting ACLs on DNS objects
